% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdry2Dgmsh.R
\name{bdry2Dgmsh}
\alias{bdry2Dgmsh}
\title{Export polygon points (vertices) to a gmsh .geo file}
\usage{
bdry2Dgmsh(x, cl = 0.1, file = "dfnBdry.geo", idStart = 1, add = FALSE)
}
\arguments{
\item{x}{2-columns table (data.frame, matrix, array). The first column are the x coordinates and the 2nd column the y coordinates of the open polygon.}

\item{cl}{Numeric scalar. Characteristic length of gmsh. It tunes up the meshing.}

\item{file}{The name of the exported .geo file.}

\item{add}{This helps to save the DFN.geo to a file}
}
\value{
If \code{add}, a file named <file>.
}
\details{
The polygonal boundary does not need to enclose all the gmsh elements; but the meshing will only be made inside the polygon. See examples.
}
\examples{
x <- cbind(x = c(0,6,6,0), y = c(0,0,3, 3))
plot(x, type='l', col = 'blue')
# write boundary to file in the current working directory
bdry2Dgmsh(x)

# Another saving-to-file example
bdry2Dgmsh(x, idStart = 11)
write('// This is another file', file = "newBdry.geo")
bdry2Dgmsh(x, file = 'newBdry.geo', add = TRUE)
}
