% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfn.R
\docType{data}
\name{dfn}
\alias{dfn}
\title{Set of line segments (Discrete Fracture Network)}
\format{A data.frame. See \link[graphics]{segments} to understand column names of 'dfn'.}
\usage{
dfn
}
\description{
Set of line segments (Discrete Fracture Network)
}
\keyword{datasets}
