% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfn2geo.R
\name{dfn2geo}
\alias{dfn2geo}
\title{Line Segments to gmsh .geo format export}
\usage{
dfn2geo(x, cl = 0.1, file = "dfn.geo", idStart = 1, add = FALSE)
}
\arguments{
\item{x}{data frame of line segments ends coordinates.}

\item{cl}{Numeric scalar. Characteristic length of gmsh. It tunes up the meshing.}

\item{file}{The name of the (possibly exported) .geo file.}

\item{add}{This helps to save the DFN.geo to a file}
}
\value{
If \code{add}, a file named <file>.
}
\details{
To mesh the .geo files in the examples, see \link{exportDfn2geo}. To view this .geo <file> in gmsh software, go to the directory, then run in the terminal \code{gmsh <fileWithExtension>}.
}
\examples{
data(dfn)
plot(x = NA, xlim= c(0,5), ylim = c(0,3))
segments(dfn$x0, dfn$y0, dfn$x1, dfn$y1, col = 'blue')
# write to file in the current working directory
dfn2geo(x = dfn)

# Another saving-to-file example
dfn2geo(x = dfn, idStart = 3)
write('// This is another file', file = "new.geo")
dfn2geo(x = dfn, file = 'new.geo', add = TRUE)
}
