% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDfn2geo.R
\name{exportDfn2geo}
\alias{exportDfn2geo}
\title{dfn to gmsh .geo file generation}
\usage{
exportDfn2geo(dfn, bdry, cl, file)
}
\arguments{
\item{dfn}{data frame of line segments ends coordinates.}

\item{bdry}{2-columns table (data.frame, matrix, array). The first column are the x coordinates and the 2nd column the y coordinates of the open polygon.}

\item{cl}{cl Numeric scalar. Characteristic length of gmsh. It tunes up the meshing.}

\item{file}{The name of the exported .geo file.}
}
\value{
A file named <file>.
}
\details{
To mesh the .geo files in the examples, go to the <file> directory, and run in the terminal \code{gmsh -2 <file>}. To view run in the terminal \code{gmsh -2 <file>.geo <file>.msh}.
}
\examples{
# Example 1
bdry <- cbind(x = c(1,6,6,0), y = c(1,0,3, 3))
data(dfn)
exportDfn2geo(dfn, bdry, cl = 0.1, file = 'dfn.geo')

# Example 2
data(synthDFN)
print(min(synthDFN$lengthCm))
bdry <- cbind(x = c(-1,401, 401, -1),
      y = c(-1,-1,  401, 401))
exportDfn2geo(dfn = synthDFN, bdry,
	cl = min(synthDFN$lengthCm), file = 'newDfn.geo')
}
