% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmshEnt.R
\name{gmshEnt}
\alias{gmshEnt}
\title{Utility to make .geo (gmsh) format}
\usage{
gmshEnt(ent = "p", id = 1, ele = 2)
}
\arguments{
\item{ent}{Character representing gmsh entities. Possible options are:
"p" = Points
"l" = Line
"ll" = Line Loop
"ps" = "Plane Surface",
"pp" = Physical Point
"pl" = Physical Line
"pl" = Physical Surface}

\item{id}{Identification number.}

\item{ele}{elements of each entity.}
}
\value{
A character vector. See examples.
}
\details{
See \link{dfn2geo} code for an example on how to write the output of this function to a file.
}
\examples{
gmshEnt(en="p", id = 1:2, ele=3:4)
gmshEnt(en="l", id = 1:2, ele=cbind(1:2, 10:11))
}
