% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervaLims.R
\name{intervaLims}
\alias{intervaLims}
\title{Vector conversion to m x 2 matrix of interval limits}
\usage{
intervaLims(x)
}
\arguments{
\item{x}{Numeric vector of increasing values (\code{x[i] < x[j], i<j}) that are the limits of the partition of an interval with limits \code{x[1]} and \code{x[n]}. Where \code{length(x)=n+1}.}
}
\value{
A m x 2 matrix with each row giving the limits of one interval.
}
\description{
Vector conversion to m x 2 matrix of interval limits
}
\details{
This code was implemented having in mind a mixture of disjoint uniform distributions.
}
\examples{
xe <- c(0,1, 4,6)
intervaLims(xe)
}
