% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthDFN.R
\docType{data}
\name{synthDFN}
\alias{synthDFN}
\title{Set of line segments (Discrete Fracture Network)}
\format{A data.frame. See \link[graphics]{segments} to understand column names of 'dfn'.}
\usage{
synthDFN
}
\description{
Set of line segments (Discrete Fracture Network)
}
\keyword{datasets}
